requirements.txt
opentelemetry-sdk
opentelemetry-api
opentelemetry-exporter-otlp-proto-grpc